Make scripts/jobs portable across platforms
Windows puts PHP executable in a different location than other platforms; so,
I added a function to retrieve the PHP executable based on whether or not I'm on
Windows. The function also ensures that a default date.timezone is provided to
it, just in case it's not set (it often isn't by default on Zend Server).

I put this function in src/functions.php. To automate autoloading, I added that
script to the "files" autoload capabilities in the composer.json, and updated
the console script, front controller, and deployment scripts to only load the
autoloader.

The Comics and GithubFeed jobs were updated to use that function to retrieve the
PHP executable.