Use Router for URI generation
This patch updates views and templates to use the RouterInterface implementation
to generate URIs, instead of using hard-coded paths throughout.

To accomplish this, I created a higher-order section, `{{#uri}}`, with a trait,
`UriTrait` that can be composed into views. The higher-order section expects a
JSON string detailing the route name, and optionally its options, which it then
uses to generate a URI. Any option values that reference template variables are
passed to the accompanying renderer prior to URI generation.

At this time, all Page implementations, the blog, and the contact pages are
generating URIs through the router.