Offline-enable website
Used https://brandonrozek.com/2015/11/service-workers/ as a guide for
setting up offline capabilities for the website. Currently, it caches
the following pages out-of-the-box:

- home page
- resume
- offline page

Additionally, it will cache other pages as you visit them for the first
time, using the following rules:

- it will cache up to 25 pages visited
- it will cache up to 10 images encountered
- it will cache any assets encountered

Local testing shows this works. A later commit will attempt to seed the
offline cache with the latest 10 blog posts as well.