Tests for Unauthorized middleware
Discovered I cannot mock `$next` with an object that delegates
`__invoke()`, as it will register the call to that method, without
letting me mock it directly. As such, created some stub methods for
creating closures to encapsulate the behavior and perform assertions:

- nextShouldNotBeCalled will `Assert::fail` if called
- nextShouldExpectAndReturn will assert that invocation arguments are the
  same as those passed when mocking, and will then return the return
  value provided.

These allowed me to test the `Unauthorized` middleware functionality.