Use `&&` to force env
I noticed error messages in the deployment logs indicating that composer was
unable to create cache directories. On inspection, it was attempting to write to
the wrong location, not the one specified in COMPOSER_HOME.

It appears that, within a script context, the following was not working:

    COMPOSER_HOME=/var/cache/composer composer ...

This patch updates that to:

    COMPOSER_HOME=/var/cache/composer && composer ...

I won't know until I test it, however.