Added error handling to build script
The build script errored, but the error did not break the deployment as it
should have; my suspicion is the hard-coded `exit 0` in the script.

This patch updates the script to wrap several calls into `if` statements, in
order to catch cases where an exit code > 0 occurs, and, when detected, exit
with 1 in order to fail the build.