Remove unsafe-inline from script-src CSP settings
Disqus was the issue, due to the recommended setup. Updated to do the
following:

- Set data attribute in the disqus_thread element with the values used
  by the script.
- Created a script that merges these data attributes into the disqus
  page object when creating the disqus configuration, and passes the key
  provided to the self-executing callback that sets up the disqus
  script.

To make this work, I had to re-order how scripts are loaded by the
template as well, and modified the blog post template to load the disqus
script.

Finally, I also removed the various slideshow embeds from the resume page. They
were poor examplars anyways, and removing them simplfies the CSP settings.