Site search!
Added lunr to the project to allow creating JS-powered, offline-capable, search
capabilities for the site.

This involved also massively modifying how the build system works, to allow
minifying JS and CSS, which also means having a mechanism for enabling
distribution versions of templates, CSS, and JS.

In the process, discovered a bunch of CSS no longer being used, and removed it.