Adds a logging listener to the `ErrorHandler`
The `ErrorHandler` allows you to attach listeners that act on the
request, response, and error. This listener uses a PSR-3 logger
(provided in this application via Monolog) in order to log errors passed
to it.

Internally, uses the Monolog `ErrorLogHandler`, which spits to the
configured `error_log`; this will make it easy to see errors in the
php-cli web server, and ensure that errors go to the docker log
aggregator when deployed.