Fixes session config and feed display
I was adding an expires header for `\.xml` files... which caused XML
files served by the application to be looked for in the filesystem
instead, and hit nginx's rules for 404 issues. Removing the `xml`
extension from expire headers fixes that issue.

Sessions were failing due to the configured `session.cookie_domain`;
this was causing failures as evidently it did not match correctly,
causing the `Cookie` header never to contain value provided by
`Set-Cookie`.