Refactors Auth code as a zend-expressive-authentication extension
Creates an `AuthenticationInterface` via `OAuth2Adapter`, which does the
work of negotiating the OAuth2 client, as well as persisting
authentication data in the session.

The ComicsPage was re-done as a vanilla Page with a session and
authentication sitting in front of it.

So far, only tested fully with the `DebugProvider`; however, the
implementation appears to work as expected regarding the OAuth2 client
workflow.