Added SSL for www.mwop.net
The SSL cert I installed was for mwop.net, so handling the redirect to
mwop.net via the same vhost would not work.

What this does is add another SSL cert for www.mwop.net, and a dedicated
server for it as well in nginx. This allows it to handle the incoming
request and redirect it to the SSL variant of mwop.net.

Also downgrades stratigility version to prevent autoloading issues,
due to incompatibilities with webimpress/http-middleware-compat.