Adds search capabilities
Adds a `search()` method to the `MapperInterface` and `PdoMapper`
implementation, to allow fulltext search of blog posts.

Adds a `SearchHandler` that passes the value of a `q` query string
argument to the `MapperInterface::search()` method in order to return a
list of responses. This is mapped to GET requests to `/search`.

Updates the javascript search functionality to use `fetch()` to fetch
items via that URL and display them.