Simplified docker stack configuration
There's no need for the network configuration.

Docker stacks have default networking that allows them to see each
other, exposing ports on the internal network. You have to configure it
to expose ports on the host network, which the configuration only does
for the caddy server - which is exactly what we want.

With these changes in place, everything is working normally again.