Provide a cache-based session adapter
Adds `Mwop\CacheSessionPersistence`, which implements
`Zend\Expressive\Session\SessionPersistenceInterface`, and uses a PSR-6
`CachePoolItemInterface` instance for the purpose of storing and
retrieving session data. PSR-6 was chosen over PSR-16, as it provides
expiry features, which are not present in PSR-16.

The application now uses a redis container in addition to caddy and
swoole. To access Redis, we use the predis library as a backend to
cache/predis-adapter.