Switch from mustache to plates
This patch switches templating from Mustache to Plates.

Because Plates handles namespaces differently, a few changes had to be
made:

- No templates could be under the default namespace. As such, I moved
  pagination below the `mwop` namespace, and called all templates in the
  `data/` directory the `data` namespace.
- You cannot have multiple paths per namespace. Since I had to override
  the 401 template from the phly-expressive-oauth2clientauthentication
  package, I had to pull in only its dependencies so I could add my own
  path to the template.

Additionally, I had to modify how the `DisplayPostMiddleware` passes
variables to the renderer. It had passed an `EntryView` as the
variables. This does not work with vanilla renderers, and, moreover,
would not allow me to call methods on it. As such, I pass it as the
variable `post` now. In the template, I provide a `uriHelper` callback
so that URL resolution will work correctly; I plan to refactor this in a
later patch.