Simplify templating logic
Removes the EntryView. Instead, a new Plates extension is created and
registered with a number of functions for creating URIs from posts,
formatting dates, and marshaling tag information.

This also led to changes in how feeds are created and posts cached. In
particular, a base URI is now created and injected into the
ServerUrlHelper to ensure absolute links are... absolute. The
ServerUrlMiddleware was also added to the pipeline for the same reason.