Make caching middleware true middleware
The functionality for displaying posts is already a handler. It's easier
to treat the `CachingMiddleware` as actual middleware, and have it
delegate to the provided handler when necessary than to maintain the
"middleware" internally.

Thus, the delegator factory for the CachingMiddleware now creates a
MiddlewarePipe, composing first the CachingMiddleware, than the result
of the `$callback`.