Migrate to Bootstrap 4
This patch migrates the site to Bootstrap 4. This required changing the
theme from Readable to Litera, as Readable is not available in
Bootswatch 4; as a result, a number of tweaks also had to be made to the
CSS to get it to match.

Part of the migration entailed moving from LESS to SASS, as Bootstrap no
longer provides LESS sources, only SASS.

Bootstrap's grid system, navbar system, and pagination system changed
dramatically, requiring a ton of changes ot make everything look okay.
In the end, I think this ends up with quite a number of improvements,
including smaller page sizes in many cases.