Succesful image built on alpine3.8
This provides a successful image built on alpine 3.8, almost halving the
size of it!

Unlike previous revisions, most extensions are added as APK packages,
speeding up builds tremendously. The one exception is the tidy
extension. This extension is built on libtidy 5.6.0 in alpine 3.8,
which, unfortunately, creates output that libxml cannot parse. As such,
this patch adds the alpine 3.6 community repository, installs
tidyhtml-dev==5.2.0-r1, and then builds the tidy extension explicitly.