Creates and registers Swoole task workers
Creates a task worker class that uses a PSR-14 message notifier to
process incoming tasks. Task start and end times are logged, as are any
errors.

Uses phly/phly-event-emitter for both the `MessageNotifierInterface` and
`ListenerProviderInterface` implementations, and wires them into the
container.

A delegator on the Swoole HTTP server adds the task worker to the
server.