Create a dedicated Redis image, and add it to the build chain
Creates a dedicated Redis image containing the configuration I need for
deployment (which includes setting a memory limit, and adding a
least-recently used expiration policy).

Adds an entry to the Makefile for creating and pushing the redis
container.

Adding/removing/changing options for the create-docker-stack command has
been painful, so this patch also refactors the create-docker-stack and
get-latest-tag commands to be symfony/console commands. Updates the
Makefile to reference the new `docker:create-stack` command.