refactor: Update to 0.7.0 of psr/event-dispatcher
This patch updates the website to use the 0.7.0 version of the
psr/event-dispatcher specification, via phly/event-dispatcher (instead
of the previous phly/event-emitter). This mainly entails:

- Updating namespaces of imported classes.
- Updating the names of several imported classes.
- Removal of event/message/task implementations; events are simply
  objects now.
- Switching from `ListenerAggregate` to `AttachableListenerProvider`,
  and thus from `on()` to `listen()`.

Additionally, because dispatchers stop execution on the first thrown
exception, much of the logic of the TaskWorker could be removed, as
there is no longer an idea of an aggregate exception.