refactor: Use events to cache blog posts
Adds a `CacheListener` that accepts a new `CacheBlogPostEvent` (which
aggregates a cache item and a response); the listener then uses a
composed cache pool to cache the item with a serialized version of the
response. This replaces usage of Swoole\Event::defer();

The idea is to have a "queuable" listener type in the `TaskWorker`
namespace, which will then decorate an event such that it creates a task
with the callable and event, allowing it to execute asynchronously. This
will allow decoupling the application code from Swoole-specific code.