refactor: compose event dispatchers, and implement queueable listeners
This patch refactors how task deferment occurs.

Previously, handlers would compose the Swoole HTTP server instance, and
call its `task()` method with a message/task/notification. The task
workers then would pass the task to the event dispatcher to handle.

THis patch inverts that relationship. Handlers now compose an event
dispatcher, and dispatch an event. The task worker now expects an
`Mwop\TaskWorker\Task` instance, which contains an `object` event and
`callable` listener; it will then call the listener with the event as an
argument.

To make a particular listener use the task workers, it can do one of the
following:

- Decorate itself in an `Mwop\TaskWorker\QueueableListener`, along with
  a reference to the HTTP server instance. When invoked, this will
  create a `Task` with the event passed to it and the composed listener,
  and pass it to the HTTP server.

- Mark itself as `Mwop\TaskWorker\ListenerShouldQueue`. If the listener
  provider is an `Mwop\TaskWorker\QueueableListenerProvider`, it will
  decorate the listener in a `QueueableListener` before attaching it.

This approach is more flexible for a variety of reasons.

First, it means that the handlers and other application code are not
aware of the environment in which they run; they simply dispatch events.

Second, this means we can have multiple listeners, some of which operate
synchronously. This can be useful when we want to otherwise alter the
event, or perform actions that should occur before a response is
returned, while still allowing listeners to defer execution to later.