refactor: Create an `Mwop\App` module
This patch grabs all remaining code and pushes it into an `App`
submodule. Additionally, it splits the handlers into an additional
`Mwop\App\Handler` namespace, the middleware into `Mwop\App\Middleware`,
and pushes the "Feed" and "Factory" subnamespaces under `Mwop\App`. All
factories are moved adjacent to the classes they produce, unless they
are for third-party services.

A ConfigProvider is created which contains dependency configuration, as
well as default configuration for the "cache",
"content-security-policy", "homepage", and "mail" services.  Any
third-party services on which any class or factory depends within the
`Mwop\App` namespace is added to the wiring as well.

Handlers are given the `Handler` suffix while middleware are given the
`Middleware` suffix, and factories are updated accordingly. The
`PageFactory` becomes `PageHandlerFactory`, and is updated to strip the
`Handler` suffix, as well as simplify its logic for identifying the
template name.

The `ResponseFactory` was removed, as it is no longer used.