refactor: move OAuth2 route registration to module
Creates the method `ConfigProvider::registerRoutes()`, which takes a
string base path and the application, which it then uses to register the
relevant OAuth2 routes.

The application routing table then creates an instance of the
`ConfigProvider` and calls that method to register the OAuth2 routes.

This approach makes it clear that the module contains routes to register
with the application, which then allows a maintainer to easily identify
_where_ that happens as well.