refactor: do not hardcode URLs in Contact module
The Contact module was using the "original messages" middleware to
hard-code URLs for the form action and redirects. This patch modifies it
to instead use the URL helper where possible. In the case of the form
action, it now uses the Plates URL helper directly in the template
instead of passing the URL as a template value.