fix: prevent caching of instagram feed indefinitely
This patch updates how how the Instagram feed is cached. When
caching it as a config file, the feed is never re-read by the
application, as the config is cached for the lifetime of the application
(i.e., until it is redeployed). As such, it needs to be cached
elsewhere.

It is now cached in the `data/instagram.feed.php` file, which is read
_every time_ the home page is hit.