fix: ensure that instagram client can handle carousels
When an item contains a carousel (i.e., multiple different images), its
`getImageVersions2()` method returns null. This patch updates the code
to check for a non-null value before attempting to pull image candidates
from it. If a null value is found, it then checks for a carousel,
looping through each of its items to look for images to return. Finally,
if neither works, it will yield from an empty array, preventing any
errors from bubbling up.