refactor: rely on `apt install` to update packages
We should likely be updating to the latest versions of packages on each install to ensure we get security updates.
`apt install` will do this if the package is already installed and a newer version exists.
As such, we need to run `apt update` before any such operations, but only once.
We can also eliminate conditionals where we were looking for whether or not a package was installed.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>