feat: use zendphp + openswoole
Updates to use ZendPHP docker image in a multi-stage build that builds assets using node, and openswoole using an intermediary zendphp container itself.

PHP runs as a non-privileged user via an s6 daemon.

To make this work, I created a patch for the mezzio-swoole library to allow it to work with openswoole, and have pinned my repo to the branch containing that patch..