refactor: unify swoole and application event dispatcher
This patch changes the application to use a single PSR-14 event dispatcher for both Swoole and the application.
Doing so allows the ability to define asynchronous tasks by wrapping them via the DeferredServiceListenerDelegator.

This means that the contacts handler now is properly doing asynchronous handling as well, and the new tweet latest blog post listener will be initially setup to do so, too.