refactor: create cron component and extract Comics module
While the approach with registering a Tick per task type works, it means a fair amount of boilerplate for any new periodic jobs.

This patch does a few things.

First, it extracts all comics-related items (handlers, console tools, periodic tasks) into a single module.
This affects each of the App and Console modules, and removes most of the `Mwop\App\PeriodicTask` functionality.

Second, it creates a new module, `Mwop\Cron`.
This module does the magic.
It takes an array of jobs, each being a map of:

- schedule => crontab expression
- event    => event class name representing the job to trigger

Events must implement `Mwop\Cron\CronEventInterface`, which defines two methods:

- `public static function forTimestamp(DateTimeInterface): self`
- `public function timestamp(): DateTimeInterface`

A delegator on the `Phly\EventDispatcher\AttachableListenerProvider` registers a `Mezzio\Swoole\Event\ServerStartEvent` listener that creates an `Mwop\Cron\Dispatcher` instance, and registers it with a `Swoole\Timer::tick` that runs every minute.

The `Dispatcher` accepts the application event dispatcher, a logger, and a Crontab instance (containing all cronjobs).
Each time it is invoked, it iterates through the Crontab and checks each item to see if it is due using dragonmantank/crontab-expression.
If any item is due, it uses the associated event's `forTimestamp()` method to create an event that it then dispatches via the application event dispatcher.

What this means is that to register a cronjob, you do the following:

- Create a `CronEventInterface` implementation.
- Create a listener for that event.
- Wire the listener to the event dispatcher (usually via a listener provider implementation); if you want to run it async, decorate the listener using mezzio-swoole `DeferredServiceListenerDelegator`.
- Add cron configuration:

  ```php
  return [
      'cron' => [
          'comics' => [
              'schedule' => '0 */3 * * *',
              'event'    => Mwop\Comics\ComicsEvent::class,
          ],
      ],
  ];
  ```

This approach ensures that there is error handling (already built into the mezzio-swoole task listener), and prevents needing to know how to register ticks with Swoole.