fix: use server->tick instead of Timer::tick
This keeps it in the server's event loop.

I'd hoped to refactor so that the listener could be pulled from the container, but I ran into the same race condition as previously: the dispatcher needs the EventDispatcher, and the listener needs the dispatcher, which leads to a race condition.
As such, I've kept the change minimal.