fix: use CurlHandler instead of default CurlMultiHandler with AWS SDK
Because the client is _stateful_, the first attempt to fetch an image sets the curl multi handle resource in the instance.
Once set, parallel requests that come in are blocked from working, as its already been initialized.

Under Swoole, non-multi-handle requests are run in coroutines, so they already benefit from parallelization.

As such, the solution is to use the CurlHandler instead of the CurlMultiHandler (which is the default for a guzzle client instance).