refactor: move retrieval as PSR-7 response to new class
Allows using the raw S3 client, which, now that I'm using the CurlHandler, returns a body stream that doesn't cause errors in Swoole.
This speads up returns by a factor of around 150%, and reduces memory usage, as it's essentially a pass-through, and because I only need to contact S3 once, instead of multiple times (once each to get the file, the MIME type, and the content length previously).