feat: working CSS and template for providing an image upload
Derived fom https://codepen.io/Scribblerockerz/pen/qdWzJw

The solution is a combination of HTML and CSS, and it allows either clicking to provide a filename, or dragging to the area.
When a file has been selected, the area receives a green checkmark.
The input is filled, and the value passed on to whatever page then processes the upload.

Additionally, I discovered that Swoole by default only supports 2MB payloads; I upped it to 48MB.