feat: upload photos
Provides a gated page that allows uploading photos to the gallery.
A new "UploadPhoto" service does the actual work, which is similar to that of the webhook payload listener, but differs in that it passes the uploaded resource directly to storage.

On successful completion, it redirects to the new page for the artwork.