feat: now page archives
"Now" page entries are now stored in my object storage.
I grab a list of them, and sort that list in reverse order by date, only grabbing those named "Y-m.md".
When /now is requested, it uses the first entry, and displays the page as "Now".
A list of previous pages is listed at the bottom, if there are any, and it links to each; these are labelled "Then".
Otherwise, you can request pages as /then/{Y-m}.