feat: adds caching middleware
The middleware can be placed in front of any route that should cache content.
It attempts to pull a response from the cache; if none is found, it then delegates to the handler, and attempts to cache the response returned by it.
If a cache hit is detected, it deserializes and returns the response.

The middleware uses the response serializer from Diactoros to serialize and deserialize from a string.

Additionally, I've added:

- Functionality to the container initialization script to clear the response cache.
- An admin handler for clearing the response cache; this handler is gated by authn/authz.

And:

- I refactored route definition of admin routes to define an array with default middleware, and then array unpacking to append the handler to that array.