refactor: use app response caching middleware instead of event system to cache blog posts
Modifies the `DisplayPostHandler` to always fetch a post from the mapper, and to use a PSR-17 response factory to create and return the response.
Route configuration changes to use the app `CacheMiddleware` with this handler.

Testing locally indicates it works, which means I can remove the events and the console tooling for clearing the cache.