fix: ensure generated content has styles that will apply
Home page feeds, comics, and search results are all either generated by tools, or injected via JS, and, as such, because tailwind could not find templates with those values, it was not retaining the styles.
The key in those situations is to put them OUTSIDE all layers; you can still use `@apply`, however.