refactor: feed item interface, value objects, and constructor
Makes `FeedItem` an interface, and creates two implementations:

- `InvalidFeedItem` is an empty version indicating the provided feed should not be injected (either due to error or due to rules)
- `PopulatedFeedItem` is a version that is fully populated and ready to inject into the feed list.

The `MapperBuilderDelegator` is a delegator on the cuyz/valinor `MapperBuilder` that adds a constructor for a `FeedItem`.
The constructor details the required items (all non-nullable, and with either strong type hints or Psalm hints), as well as optional items (nullable).
It then decides if it matches known rules, returning an `InvalidFeedItem` if it does not, and otherwise attempting to populate a `PopulatedFeedItem` (in at least one case, it will throw an exception for cuyz/valinor to catch for invalid data).

The config provider ensures that expected configuration for mapping favicons is present.