fix: ensure that we can parse feed payloads
Discovered that I needed to setup interface inference in cuyz/valinor, which in turn led to the discovery that the builder is immutable, and its methods return **new** instances.
This also means that the logic is simpler for the constructors.

One issue: `HomepagePostsList` evidently used the `fromArray()` method; this will need to be refactored.