refactor: use cuyz/valinor for Photo webhook payloads
Makes Photo a minimalist value object, removing the `toArray()` functionality as well as the `transformStringDateTime()` functionality.
Creates an `Mwop\Art\MapperBuilderDelegator` class that adds a constructor for the `Photo` entity that ensures that the source_url and created_at strings are non-empty-strings, and then brings in the `transformStringDateTime()` method for purposes of created a `DateTimeImmutable` instance from the `created_at` value prior to instantiating the `Photo.`
The PdoPhotoMapper now creates `DateTimeImmutable` instances for the `createdAt` argument of a `Photo` before instantiating it.

Tested locally with an IG payload; works correctly.