feat: allow posting Art to Mastodon
Creates a `PostToMastodon` class in the `Mwop\Art` namespace that will:

- Find a photo by filename, and post it to Mastodon
- Find the most recent photo posted, and post it to Mastodon

In doing this, I discovered that I did NOT have a PSR-18 client service setup; this is now created.
I also discovered I needed a way to generate a `Media` instance from an image name, and that the `Media` instance should allow either resources OR PSR-7 streams.

[optional footer]

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>