refactor: working php-fpm + supervisord workers setup
Refactors the DEVELOPMENT docker-compose environment such that it now works as follows:

- redis node
- worker node running supervisord, which manages task and cron workers (task workers use redis)
- nginx node, sitting in front of:
  - php-fpm node

The supervisord and nginx nodes run as the zendphp user, and use a stdout logger that Docker then picks up.

The nginx configuration adapts the old swoole configuration for handling MIME types and setting cache-control/etag headers.
That node also gets the generated CSS and JS, as well as all files from the `public/` tree.

The php-fpm and worker nodes share a volume, which allows the worker node to do all the initial setup tasks (generating the blog, grabbing the photo database, grabbing initial social media and feed information, etc.).

Most everything works, except for site search.
This could be a CSP issue, or it could be due to the lack of generated search terms; I still need to investigate that part.