feat: healthchecks for all services, plus production config
Sets up health checks for all services.
For supervisor, this meant configuring the RPC service on a local socket so that supervisorctl could provide status.
For php-fpm, I found a script by Renato that provides healthchecks.

Adapts the production configuration to follow that of the development config, with the notable exception that nginx will map host 9001 to container 80, to allow the existing reverse proxy to work.