refactor: for production, create containers that map in all files
For each of the nginx, worker, and php images, this patch adds template dockerfile content that during deployment will be used to create "production" dockerfiles.
The templated pieces use COPY to add files from the application to the container, and, in the case of the worker and php images, also runs code to install production application dependencies.