fix: do not use getClientMediaType()
Since UploadedFileInterface::getClientMediaType can lead to errors, and, with the switch to php-fpm, I've run into that error, this patch adds a `getMediaType()` method in `UploadPhoto` that determines it using `mime_content_type()` on the uploaded file stream URI metadata.

This value is then passed to `PhotoStorage::fromUploadedFile()`, and used when present instead of `getClientMediaType()`.

Also, `UploadPhoto::generateError()` was missing one value when generating the templated log message, which would lead to a 500 error.