fix: ensure that cron queues are created at launch
Because ZendHQ does not yet work on the CLI SAPI, I have to resort to a few gymnastics to ensure my cron queues are created as early as possible.

To that end, I've created a "health" handler that returns a 204, and set that up as the nginx healthcheck.
This ensures that the application is hit within 30 seconds of launch, which then ensures the queues are created.